requirements.txt
requests
termcolor
prettytable
ollama
flask
google-api-python-client
google-auth
gunicorn
